Drop -U ar flag from Makefile
In some (most?) distributions these days, binutils is built with
"deterministic mode" as the default. This means that ar will omit file
metadata information from the archive, so that its contents are
dependent only on the contents of its members and nothing else. This is
great for packaging because it entail reliable checksumming, but also
breaks Make's ability to update an archive incrementally by inspecting
the timestamps that would have been otherwise embedded in the archive.

The -U flag disables deterministic mode, and is typically to used to
recover this GNU Make feature. However, we have three reasons to exclude
it: (1) this breaks compatibility with macOS, whose ar distribution
doesn't appear to even recognize this flag; (2) this archive is tiny,
and rebuilding the entire thing takes very little time; (3) our Makefile
isn't even currently written to make use of this Make feature. (In fact,
the way it written now, all lib oject targets depend on each other,
which is a bug I will fix in a later commit.)