Work in progress
Notes to self:

I realize that the abstraction I'm unsuccessfully recreating here is
some ssm_unique_owner(), which just checks whether the ref count is 1.
If so, then the compiler is free to modify that value in-place, like
reuse. Otherwise, the compiler must clone it.

This applies not only to closures but also other object kinds too.

For closures, perhaps all we need is some kind of clone method and some
kind of apply method.